refactor(takt): extract shared parseTaktConfig helper
Address PR review (mid DRY): the identical parseTaktConfig YAML reader was duplicated in takt-mcp.ts and takt-permissions.ts. Extract it to src/features/shared/takt-config.ts (with tests) so future Takt features that read .takt/config.yaml reuse one implementation.
